requirements.txt
Django>=5.0,<6.0
openai>=1.0.0
django-allauth>=0.65.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
